python_web/cart/pyproject.toml
[project]
name = "cart"
version = "0.1.0"
description = "Cart microservice"
requires-python = "==3.12.*"
dependencies = [
    "backend-common",
    "fastapi>=0.129.0",
    "asyncpg>=0.31.0",
    "pydantic-settings>=2.13.0",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.41.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "aiokafka>=0.10.0",
    "alembic>=1.15.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.25.0",
    "aiosqlite>=0.21.0",
    "coverage>=7.13.5",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv.sources]
backend-common = { path = "../shared", editable = true }

[dependency-groups]
dev = [
    "pytest-cov>=7.1.0",
    "ruff>=0.15.8",
]
